secrets/filters.py: partial restore of 3774229978e046a7d5781e2b83a63866f2cccab9
Unifying these extra models with the ones in the new `enums.py` proved
challenging, due to the immutability of enums and the divergent
requirements of the classes. The sustainable way to do this would
involve changing the way these models are used, which is out of scope of
this PR and should be tackled at a future date.