"Downgrade" version number
The scope of the project has exploded since the original 1.0. We currently don't feel like it's at a stable 1.0 for production use other than our own company because of missing docs and tests. Having a 1.0 that is an Alpha feels wrong, so we'll go back to 0.X releases for now.