merge build steps into one so static files are included in pypi release
By splitting the build process in two, the steps for building the
webpack files and python distribution were separated such that the
webpack files were not available when we built the wheel. This meant
that they were not included when installing via pip (or uv), resulting
in a broken installation.
For added security, added a smoke test to avoid pushing broken releases,
and a dry-run for the publishing for manual runs.